\name{bn2alk}
\alias{bn2alk}
\title{
gets an alk from a given bn
}
\description{
(bn) extracts the alk object associated to the node number \code{nk}
}
\usage{
bn2alk(bn, nk)
}
\arguments{
\item{bn}{bn object.}
\item{nk}{number of the node.}
}
\details{
To allow the produced /alk/ to be used with other /bn/ it is of type
non completed.
}
\value{
a non completed /alk/
}
