\name{rebastaba}
\alias{rebastaba}
\title{
draws some Bayesian networks for illustration
}
\description{
(mi) This function draws some Bayesian networks using rebastaba
spelling to serve as a cover figure for the manual
}
\usage{
rebastaba()
}
\examples{
 g4n3k("RESET");
 rebastaba();
}
\arguments{
}
\value{
nothing but a plot is performed
}
