\name{valid8bn}
\alias{valid8bn}
\title{
checks a /bn/
}
\description{
(bn) This function checks /bn/ objects
}
\usage{
valid8bn(object)
}
\arguments{
\item{object}{The bn object to be validated.}
}
\details{
It is the validity method for /bn/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
