\name{valid8win}
\alias{valid8win}
\title{
checks a /win/
}
\description{
(ba) This function checks /win/ objects
}
\usage{
valid8win(object)
}
\arguments{
\item{object}{The win object to be validated.}
}
\details{
See the code or the associated documentation.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
