\name{variables8win}
\alias{variables8win}
\title{
returns all variables involved in a /win/
}
\description{
(ba) finds the variable indicated by a \code{win}.
}
\usage{
variables8win(win)
}
\examples{
 rebastaba3k("RESET"); # needed only for R checking, to be forgotten
 variables8win(rebastaba.win1);
}
\arguments{
\item{win}{win object}
}
\details{
No check is performed.
}
\value{
the names of the variables.
}
\keyword{misc}
